Modify some of the tests to remove dependency onv1
